refactor: move tr_makeFilename() to tr_torrent_metainfo
